requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==3.1.3.
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
lazy-object-proxy==1.4.3
nginx==0.0.1
packaging==20.3
pillow==7.0.0
psycopg2-binary==2.8.5
py==1.8.1
PyJWT==1.7.1
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.23.0
sqlparse==0.3.0
sorl-thumbnail==12.6.3           
urllib3==1.25.6           
wcwidth==0.1.8            
zipp==2.2.0